requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
Django==5.1.7
Flask==3.1.0
flask-cors==6.0.1
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.4.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
oscrypto==1.3.0
pillow==11.0.0
protobuf==3.20.3
pycparser==2.22
PyGetWindow==0.0.9
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyMsgBox==1.0.9
pypdf==5.6.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
pytweening==1.2.0
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
scipy==1.14.1
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
wavio==0.0.9
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
